Big refactor using classes; add controls
Replace the procedural app structure with an app class: a base_app which
is supposed to be generic and reusable (although that goal might not yet
be met) and a browser_app which inherits from it which is specific to
this project.

Windows are now implemented as window classes: a generic base_window and
a specific browser_window. Each browser_window has a browser_tab, with
the intention of supporting multiple tabs within each browser window,
but this is completely unimplemented so far.

The browser window now has scroll bars, a header, some buttons, and an
address bar, none of which do anything yet but at least they draw and
move and resize as they should when the window is resized. These are
implemented as standard Mac OS controls wrapped by a base_control class.
There's also a root_control class for the new root control concept
introduced in the Appearance Manager.

Replace types like short and unsigned long with their modern and
more-explicit counterparts like int16_t and uint32_t. Replace nil with
nullptr. Replace C-style casts with reinterpret_cast.

There are some regressions in the drawing of the web view. It doesn't
resize with the window anymore, it's not positioned correctly (it's
partially under the header), and it doesn't always draw when it needs to
(such as when the window is first created or when it's zoomed). These
deficiencies will be corrected later.