Give each quickdraw_font its own port
Create a new QuickDraw GrafPort when a quickdraw_font is constructed and
delete it when it is destructed.

quickdraw_fonts are constructed when the HTML document is parsed and we
need to return the font's metrics at that time, and we need to set the
port's font to do that since GetFontMetrics operates on the current
port's font. Similarly, text widths are computed when the document is
rendered and the results of TextWidth pertain to the current port's
font. Neither at parsing time nor at rendering time can we expect the
caller to have set the port; we can only expect that at draw time.

Previously we were changing the current port's font which, since the app
had not set it differently, was the window manager's port which caused
window titles to draw with the wrong width (because the width was
calculated for the wrong font).

We can't just save the current font, set our font, get the metrics, and
set it back to the previous font because there are no functions for
getting the current font from a port. We could dive directly into the
GrafPort struct and save and restore the txFont, txFace and txSize
fields, but Inside Macintosh cautions against setting those fields
directly, advising instead to call TextFont, TextFace, and TextSize,
implying that those functions might do more than just set those fields,
therefore I am doubtful that saving and restoring those fields would be
sufficient.